requirements.txt
streamlit>=1.28.0
python-docx>=1.0.1
convertapi>=1.5.0
pandas>=2.2.0
openpyxl>=3.1.0
pytest>=7.0.0
